feat: 댓글기능 완료

+++ src/main/java/net/likelion/bebc25/recipe/reply/controller/ReplyController.java
@@ -1,6 +1,9 @@
 package net.likelion.bebc25.recipe.reply.controller;
 
+import jakarta.servlet.http.HttpSession;
+import net.likelion.bebc25.recipe.member.dto.MemberDto;
 import net.likelion.bebc25.recipe.reply.dto.RequestDTO;
+import net.likelion.bebc25.recipe.reply.dto.ResponseDTO;
 import net.likelion.bebc25.recipe.reply.service.ReplyService;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -21,32 +24,46 @@ public class ReplyController {
 
     @PostMapping("/write")
     public String writeReply(@RequestParam("boardType") String boardType,
-                             @ModelAttribute RequestDTO requestDTO) {
+                             @ModelAttribute RequestDTO requestDTO,
+                             HttpSession session) {
 
-        requestDTO.setMemberId(1L);
-        // 댓글 저장
+        MemberDto loginMember = (MemberDto) session.getAttribute("loginMember");
+
+        if (loginMember == null) {
+            return "redirect:/login";
+        }
+
+        requestDTO.setMemberId(loginMember.getId());
+
         replyService.writeReply(requestDTO);
 
-        //어느 게시판이냐에 따라 리다이렉트 경로를 다르게 설정!
         if ("tip".equals(boardType)) {
-            // 팁 게시판 상세로 이동
             return "redirect:/tip/detail?id=" + requestDTO.getPostId();
         } else {
-            // 레시피 게시판 상세로 이동
             return "redirect:/recipe/detail?id=" + requestDTO.getPostId();
         }
     }
 
-
     @PostMapping("/delete")
-    public String deleteReply(@RequestParam("id") Long id,
-                              @RequestParam("postId") Long postId,
-                              @RequestParam("boardType") String boardType) {
+    public String deleteReply(@RequestParam("id") int id,
+                              @RequestParam("postId") int postId,
+                              @RequestParam("boardType") String boardType,
+                              HttpSession session) { //HttpSession 파라미터 추가
 
-        // 1. 댓글 삭제
-        replyService.deleteReply(id);
+        MemberDto loginMember = (MemberDto) session.getAttribute("loginMember");
 
-        // 2. 어느 게시판이었느냐에 따라 상세 페이지로 리다이렉트
+        if (loginMember == null) {
+            return "redirect:/login";
+        }
+
+        ResponseDTO reply = replyService.findById(id);
+
+        if (reply != null && reply.getMemberId() == loginMember.getId()) {
+            replyService.deleteReply(id);
+        } else {
+            System.out.println(">>> [삭제 실패] 본인이 작성한 댓글만 삭제할 수 있습니다.");
+        }
+
         if ("tip".equals(boardType)) {
             return "redirect:/tip/detail?id=" + postId;
         } else {
